Blast and blat parser changes
I created an abstract class called “databaseSearchOutput.java” with the
following sub classes: BlatOutput and BlastOutput. I did this because
in order to use blat i had to use the eggNog database as a query (since
it is a protein database) and the reads as a target (the reads are
DNA).  this switches the target and query information so I wanted to
make a separate class called “BlatOutput.java” since this is so close
to “BlastOutput.java” I made them subclasses of a bigger class called
“databaseSearchOutput.java”.